requirements.txt
annotated-types==0.6.0
anyio==3.7.1
bcrypt==4.2.0
certifi==2023.7.22
cffi==1.17.1
charset-normalizer==3.3.0
click==8.1.7
cryptography==43.0.1
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
exceptiongroup==1.1.3
fastapi==0.103.2
greenlet==3.1.1
h11==0.14.0
idna==3.4
passlib==1.7.4
psycopg2-binary==2.9.9
pyasn1==0.6.1
pycparser==2.22
pydantic==2.4.2
pydantic_core==2.10.1
python-jose==3.3.0
python-multipart==0.0.12
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.35
starlette==0.27.0
typing_extensions==4.8.0
urllib3==2.0.6
uvicorn==0.23.2
